convert DeckCallout to TS
