+++ src/interfaces/IDeckCalloutProps.ts
@@ -1,0 +1,9 @@
+/* custom stuff */
+// import AppContext from "../classes/AppContext";
+
+export interface IDeckCalloutProps {
+  player: any;
+  isDeckCalloutVisible: boolean;
+  onHideCallout: () => void;
+  target: string;
+}
